Moved formatting into separate modules
